pyproject.toml
[tool.poetry]
name = "jira-actions-lock"
version = "0.1.0"
description = ""
authors = ["Brandon White <bwhite@infoxchange.org>"]

[tool.poetry.scripts]
jira_check = "jira_actions_lock.__main__:main"

[tool.poetry.dependencies]
python = "^3.9.7"
click = "^8.1.3"
requests = "^2.28.1"
GitPython = "^3.1.27"

[tool.poetry.dev-dependencies]
pre-commit = "^2.12"
pylint = "*"
mypy = "^0.971"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

